Add files via upload
I have made a repository of this project to help others to make similar projects. I have uploaded program files and other images like of power supply and components used and all other necessary steps taken to complete the project. i have alo shown the use of android app in my project and how we can control our home appliances with the help of that. Please go through the repository for more details